(3.0.7) 오타수정

+++ SV_Simulator_v1/interface/interface.cpp
@@ -33,7 +33,7 @@ int SV_Sim::Run()
 		if (SV_Sim::simState == SimState::Pause)
 			SV_Sim::Pause();
 
-		Sleep(waitTime);
+		Sleep(static_cast<DWORD>(waitTime));
 
 		if (SV_Sim::simState == SimState::WaitEnd)
 			break;
